XSD: Add styling to root element, add license to array variations.
